Assignment requirements done with error: signal handler, sleep, zombie protection. malloc() error did not investigated.

+++ hw2_2024/hw2.c
@@ -9,6 +9,7 @@
 #include <time.h>
 #include "common.h"
 #include <sys/wait.h>
+#include <signal.h>
 
 #define FIRST_CHILD_FIFO "/tmp/firstFifo"
 #define SEC_CHILD_FIFO "/tmp/secFifo"
@@ -23,70 +24,19 @@ int addValuesInInteger(int* convertedArray, int size){
 }
 
 void intoStringWithDelim(int sum, char* buffer, const char delim){
-    sprintf(buffer, "%d", sum);
-    
+    int num_written  = sprintf(buffer, "%d", sum);
+    printf("WRITTEN SIZE : %d\n", num_written);
     int len = 0;
-    while (buffer[len] != '\0') {
-        len++;
-    }
-        buffer[len] = delim;
-    buffer[len + 1] = '\0';
+    buffer[num_written] = delim;
+    buffer[num_written + 1] = '\0';
 }
-// char* readOneByOne(const char* firstChildFifo){
-    
-//     int numBytesRead = 0;
-//     int childBufferCounter = 0;
-//     char* buffer = (char*)malloc(sizeof(char)*6);
-//     // char buffer[256];
-//     memset(buffer, 0, 6);
 
-//     int fd = open(firstChildFifo, O_RDONLY);
-//     if(fd == -1){
-//         perror("open failed");
-//         free(buffer);
-//         exit(EXIT_FAILURE);
-//     }
-//     for(;;)
-//     {
-//         char c; 
-//         char log[256];
-
-//         memset(log, 0, sizeof(log));
-//         fflush(stdout);
-//         numBytesRead = read(fd, &c, sizeof(c));
-
-//         if(numBytesRead == 0){
-            
-//             buffer[childBufferCounter] = '\0';
-//             break;
-//         }
-//         else if(numBytesRead == -1){
-//             perror("read");
-//             free(buffer);
-//             return NULL;
-//         }
-//         else{
-//             if(childBufferCounter+1 % 6 == 0){
-//                 buffer = (char *)realloc(buffer, sizeof(char)*(childBufferCounter*2)); 
-//             }
-//             buffer[childBufferCounter++] = c;
-//         }
-//     }
-//     if(close(fd) == -1){
-//         perror("close");
-//         free(buffer);
-//         return NULL;
-//     }
-
-
-//     write(STDOUT_FILENO, buffer, strlen(buffer));
-//     return buffer;
-// }
 long int multiplyValuesInInteger(int* convertedArray, int size){
     long int result = 1;
     for (int i = 0; i < size; ++i) {
         result *= convertedArray[i];
     }
+    printf("HIIII\n");
     return result;
 }
 int multiplyOperation(int* convertedArray, int count){
@@ -111,39 +61,6 @@ int writeToFirstFifo(const char* str, int fd){
     return 1;
 }
 
-// void parseAndCalculate(const char* input) {
-//     int multiplicationResult = 1;
-//     int additionResult = 0;
-//     char* token;
-//     char* tempInput = strdup(input); // Make a copy of the input string to tokenize
-
-//     // Parse the input string
-//     token = strtok(tempInput, ",");
-//     while (token != NULL) {
-//         if (strcmp(token, "multiply") == 0) {
-//             // Perform multiplication
-//             token = strtok(NULL, ",");
-//             while (token != NULL && token[0] != '!') {
-//                 printf("multiply token: %s\n",token);
-//                 multiplicationResult *= atoi(token);
-//                 token = strtok(NULL, ",");
-//             }
-//         } else {
-//             // Perform addition
-//             additionResult += atoi(token);
-//             token = strtok(NULL, ",");
-//         }
-//     }
-
-//     // Free allocated memory
-//     free(tempInput);
-
-//     // Print the results
-//     printf("Multiplication result: %d\n", multiplicationResult);
-//     printf("Addition result: %d\n", additionResult);
-//     printf("Total result: %d\n", multiplicationResult + additionResult);
-// }
-
 int parseSecondfifo(const char* buffer, const char delim){
     char** splitted;
     int* convertedArray;
@@ -151,33 +68,46 @@ int parseSecondfifo(const char* buffer, const char delim){
     int flag = 0;
     char* convertedArrayStr= NULL;
     int count = countHowManyElementsWillExtract(buffer, delim);
-    // printf("ARRAY COUNT %d \n", count);
-    splitted = (char**)malloc(sizeof(char*)*count);
-    convertedArrayStr = splitStringIntoArray_Custom(buffer, splitted,convertedArrayStr);
+    // printf("Count of elements: %d\n", count);
+    splitted = (char**)calloc(count, sizeof(char*));
+    // printf("\n\nADKSADKLASDLK successfuly done malloc 157hw2\n");
 
-    convertedArray = (int*)malloc(sizeof(int)*count);
+    convertedArrayStr = splitStringIntoArray_Custom(buffer, splitted,convertedArrayStr);
+    convertedArray = (int*)calloc(count, sizeof(int));
+    // printf("successfuly done malloc 161hw2\n");
     for(int i = 0; i < count-1; i++){
-        if(strncmp(splitted[count-1], "multiply", 8) == 0){
+        if(strncmp(splitted[i], "multiply", 8) == 0){
             flag = 1;
         }
-        // printf("ss necwcc: %d\n",strlen( splitted[i]));
+        printf("ss necwcc: %d\n",strlen( splitted[i]));
         convertedArray[i] = convertSingleStringToInteger(splitted[i]);
     }
+    if(strncmp(splitted[count-1], "multiply", 8) == 0){
+        flag = 1;
+    }
+    // printf("HIIII0 %d\n", flag);
     if(flag == 0){
-        free(convertedArrayStr);
-        free(convertedArray);
-        for(int i = 0; i < count+1; i++){
-            free(splitted[i]);
-        }
-        free(splitted);
+        // free(convertedArrayStr);
+        // free(convertedArray);
+        // for(int i = 0; i < count+1; i++){
+        //     free(splitted[i]);
+        // }
+        // free(splitted);
         return -1;
     }
+    // printf("HIIII2\n");
     // for(int i = 0; i < count-1; i++){
 
     //     printf("sas necwcc: %d\n",convertedArray[i]);
     // }
+    // printf("HIIII3\n");
     long int res =  multiplyValuesInInteger(convertedArray, count-1);
+    // printf("HIIII4\n");
+    // for(int i = 0; i < count; i++){
+    //     printf("convertedArray : %d \n", convertedArray[i]);
+    // }
     int addRes = strtol(convertedArrayStr, &endptr, 10);
+    // printf("HIIII5\n");
     printf("Multiplication result: %ld\n", res);
     printf("Addition result: %d\n", addRes);
     printf("Total result: %ld\n", res + addRes);
@@ -187,6 +117,7 @@ int parseSecondfifo(const char* buffer, const char delim){
     }
     free(convertedArrayStr);
     free(convertedArray);
+
     free(splitted);
     return res+addRes;
 }
@@ -197,7 +128,10 @@ int parseFifo(const char* buffer, const char delim){
     int* convertedArray;
     int count = countHowManyElementsWillExtract(buffer, delim);
     splitted = (char**)malloc(sizeof(char*)*count);
+    // printf("XXXXXX-------count: %d\n", count);
+    // printf("successfuly done malloc 212\n");
     convertedArray = (int*)malloc(sizeof(int)*count);
+    // printf("successfuly done malloc 215\n");
 
     splitStringIntoArray_I(buffer, delim, splitted, count);
     convertStringArrayToInteger(splitted, count, convertedArray);
@@ -215,13 +149,39 @@ int parseFifo(const char* buffer, const char delim){
     free(convertedArray);
     return sum;
 }
+static int spawnedProcess=0;
+static int signalCnt;
+static void sigchldHandler(int sign){
+    int status, savedErrno;
+    int childpid;
+
+    savedErrno = errno;
+    while((childpid = waitpid(-1, &status, WNOHANG)) > 0){
+        printf("handler: Reaped child %ld\n", (long) childpid);
+        signalCnt++;
+    }
+    if (childpid == -1 && errno != ECHILD)
+        perror("waitpid");
+    if(sign == SIGINT){
+        printf("SIGINT handled... Your operation will done after process done\n");
+    }
+
+    errno = savedErrno;
+}
 
 
 int main(int argc, char **argv)
 {
+    struct sigaction sa;
+    sigset_t prevMask, blockMask;
+    int sig;
     int bufferLen = 256;
     char buffer[bufferLen];
     char log_buffer[bufferLen];
+    int sec_fifo_fd;
+    int status1, status2;
+    int logSize = 0;
+
     // char err[256]; 
     
     if(argc != 2){
@@ -232,8 +192,8 @@ int main(int argc, char **argv)
     }
 
     int argument = atoi(argv[1]);
-    int sizeofprint = snprintf(buffer, sizeof(buffer), "%d", argument);
-    write(STDOUT_FILENO, buffer, sizeofprint);
+    // int sizeofprint = snprintf(buffer, sizeof(buffer), "%d", argument);
+    // write(STDOUT_FILENO, buffer, sizeofprint);
 
     if(argument < 0){
         char err[256] = "Argument must be greater than 0.\n";
@@ -243,8 +203,26 @@ int main(int argc, char **argv)
     }
 
     // fflush(STDOUT_FILENO);
+    
+    sigemptyset(&blockMask);
+    sigaddset(&blockMask, SIGINT);
+    // sigemptyset(&sa.sa_mask);
+    // if(sigaction(SIGINT, &blockMask, NULL) == -1 ){
+    //     exit(EXIT_FAILURE);
+    // }
 
+    if(sigprocmask(SIG_BLOCK, &blockMask, &prevMask) == -1){
+        exit(EXIT_FAILURE);
+    }
+    sigemptyset(&sa.sa_mask);
+    sa.sa_flags = 0;
+    sa.sa_handler = sigchldHandler;
+    
     //EEXIST: eger zaten boyle bir dosya ismi var ise
+    if(sigaction(SIGCHLD, &sa, NULL) == -1 ){
+        exit(EXIT_FAILURE);
+    }
+
     if(mkfifo(FIRST_CHILD_FIFO, 0666) == -1){
         perror("mkfifo failed\n");
         unlink(FIRST_CHILD_FIFO);
@@ -252,19 +230,19 @@ int main(int argc, char **argv)
     }
 
     if(mkfifo(SEC_CHILD_FIFO, 0666) == -1){
-        // char err[256]= "mkfifo failed\n";
-        // write(STDOUT_FILENO, err, strlen(err));
-        // memset(err, 0, sizeof(err));
         perror("second mkfifo failed");
         unlink(FIRST_CHILD_FIFO);
         unlink(SEC_CHILD_FIFO);
         return -1;
     }
 
+
+
     int pid = fork();
     switch(pid)
     {
         case 0 :
+        spawnedProcess++;
             sleep(1);
             // int childBufferCounter = 0;
             int childPid = getpid();//error check
@@ -274,6 +252,7 @@ int main(int argc, char **argv)
             memset(log_buffer, 0, log_written);
 
             childBuf = readOneByOne(FIRST_CHILD_FIFO);
+            // printf("\n\n\nParse Fifoya geldim \n");
             int sum = parseFifo(childBuf, ',');
 
             int numBytes = write(STDOUT_FILENO, childBuf, strlen(childBuf));
@@ -285,9 +264,11 @@ int main(int argc, char **argv)
             }
             memset(childBuf, 0, strlen(childBuf));
             intoStringWithDelim(sum, childBuf, '!');
+            numBytes = sprintf(log_buffer, "\nSum (exlamation ended): %s", childBuf);
+            write(STDOUT_FILENO, log_buffer, numBytes);
             // free(childBuf);
 
-            //open wronly fifo2
+            // open wronly fifo2
             int fifo2 = open(SEC_CHILD_FIFO, O_WRONLY);
             if(fifo2 == -1){
                 perror("open failed\n");
@@ -295,6 +276,7 @@ int main(int argc, char **argv)
                 free(childBuf);
                 exit(EXIT_FAILURE);
             }
+            
 
             numBytes = write(fifo2, childBuf, strlen(childBuf));
             if(numBytes == -1){
@@ -304,7 +286,6 @@ int main(int argc, char **argv)
                 free(childBuf);
                 exit(EXIT_FAILURE);
             }
-
             if(close(fifo2) == -1){
                 perror("close failed");
                 memset(childBuf, 0, strlen(childBuf));
@@ -331,15 +312,12 @@ int main(int argc, char **argv)
                 randomIntegerArray[i] = r;
             }
 
-            // for(int i = 0; i < argument; i++){
-            //     printf("%d,", randomIntegerArray[i]);
-            // }
             char* str = convertIntegerToString(randomIntegerArray, ',', argument); //you can change
             char logBuffer[bufferLen]; 
-            int logLen = sprintf(logBuffer, "Random values(ends with comma): %s\n",str);
+            int logLen = sprintf(logBuffer, "Random values generated (ends with comma): %s\n",str);
             write(STDOUT_FILENO, logBuffer, logLen);//error
             memset(logBuffer, 0, logLen);
-            //randomu burda uretmek gerek en basta uretirsen child erisebilir 
+
             int firstFifoFd = open(FIRST_CHILD_FIFO, O_WRONLY);
             if(firstFifoFd == -1 && errno == ENXIO){
                 perror("open failed");
@@ -369,36 +347,41 @@ int main(int argc, char **argv)
                 break;
                 //fail oldugunda error code ver
             }
-            //fork
+            // //fork
             int secPid = fork();
             switch(secPid){
                 case 0:
                 //child2
+                spawnedProcess++;
                     sleep(1);
                     // char c; 
                     char* childBuff = NULL; 
-                    int fifo = open(SEC_CHILD_FIFO, O_RDONLY);
-                    if(fifo == -1){
-                        perror("open failed");
-                        free(childBuff);
-                        exit(EXIT_FAILURE);
-                    }
-
-                    childBuff = readOneByOne(SEC_CHILD_FIFO);
-                    printf("I Read From Fifo : \n");
+                    int pid2= getpid();
+                    int log_written = sprintf(log_buffer, "Entered Child Process 2 : %d\n", pid2);
+                    write(STDOUT_FILENO, &log_buffer, log_written);
+                    memset(log_buffer, 0, log_written);
+                    // int fifo = open(SEC_CHILD_FIFO, O_RDONLY);
+                    // if(fifo == -1){
+                    //     perror("open failed");
+                    //     // free(childBuff);
+                    //     exit(EXIT_FAILURE);
+                    // }
+                    printf("asdasdas\n");
+                    childBuff = readOneByOne_Custom(SEC_CHILD_FIFO, ',', '!');
+                    printf("I Read to From Fifo : \n");
                     int numBytesSecChild = write(STDOUT_FILENO, childBuff, strlen(childBuff));
                     if(numBytesSecChild == -1){
                         perror("write failed (parent to child2)\n");
                         free(childBuff);
-                        close(fifo);
+                        // close(fifo);
                         exit(EXIT_FAILURE);
                     }
 
-                    if(close(fifo) == -1){
-                        perror("close failed in fifo 2\n");
-                        free(childBuff);
-                        exit(EXIT_FAILURE);
-                    }
+                    // if(close(fifo) == -1){
+                    //     perror("close failed in fifo 2\n");
+                    //     free(childBuff);
+                    //     exit(EXIT_FAILURE);
+                    // }
 
                     parseSecondfifo(childBuff, ',');
                     
@@ -414,13 +397,14 @@ int main(int argc, char **argv)
                     str = (char *)realloc(str, sizeof(char)*(strlen(str)+10));
                     char multip[10] = "multiply,";
                     int sizeBuffer = strlen(str);
+                    printf("size buffer , str %s, strlen(str) %d\n", str, sizeBuffer);
                     
                     for(int i =0; i < 9; i++){
                         str[sizeBuffer+i] = multip[i];
                     }
                     str[sizeBuffer+9] = '\0';
 
-                    printf("I Write to fifo %s\n", str);
+                    printf("xI Write to fifo %s\n", str);
                     int fdSecFifo = open(SEC_CHILD_FIFO, O_WRONLY);
                     if(fdSecFifo == -1){
                         perror("open failed in (fifo2) \n");
@@ -433,7 +417,7 @@ int main(int argc, char **argv)
                         free(str);
                         break;
                     }
-
+                    sec_fifo_fd = fdSecFifo;
                     if(close(fdSecFifo) == -1){
                         perror("close fifo failed\n");
                         free(str);
@@ -449,9 +433,50 @@ int main(int argc, char **argv)
             //sleep 
             //waitpid koy
             //wronly fifo2
-            wait(NULL);
-            printf("child exited\n");
+            int numDead= 0;
+            // for(;;){
+            //     int childpid = wait(NULL);
+            //     if(childpid == -1){
+
+            //         if(errno == ECHILD){
+            //             printf("No more children BYE \n");
+            //         }
+            //         else{
+            //             perror("wait");
+            //             exit(0);
+            //         }
+            //     }
+            //     numDead++;
+            //     printf("child exited %d\n", pid);
+            // }
+            // waitpid(pid, &status1,0);
+            // waitpid(secPid, &status2, 0);
+            // if(sigprocmask(SIG_SETMASK, &prevMask, NULL) == -1){
+            //     exit(EXIT_FAILURE);
+            // }
+            sigemptyset(&prevMask);
+            for(;;) 
+            {
+                
+                logSize = sprintf(log_buffer, "Proceeding...\n");
+                write(STDOUT_FILENO, log_buffer, logSize);
+                memset(log_buffer, 0, logSize);
+                if(signalCnt == spawnedProcess){
+                    break;
+                }
+                if (sigsuspend(&prevMask) == -1 && errno != EINTR)
+                    perror("sigsuspend");
+                sleep(1);
+            }
+
+            printf("All childs are dead... Terminating...Bye\n");
             free(str);
+            
+            // if(close(sec_fifo_fd) == -1){
+            //     perror("close fifo failed\n");
+            //     free(str);
+            //     break;
+            // }
 
         break;
     }
